+++ src/pipeline_iterators.rs
@@ -1,7 +1,7 @@
 use super::headers::Headers;
 use crate::target::Target;
 use crate::transform::{compute_hash, Transform};
-use crate::{Error, Pipeline, PipelineIter, Row, RowResult};
+use crate::{Error, ErrorKind, Pipeline, PipelineIter, Row, RowResult};
 use linked_hash_map::{Entry, LinkedHashMap};
 
 pub struct PipelinesChain<'a, P> {
@@ -17,32 +17,38 @@ where
 	type Item = RowResult;
 
 	fn next(&mut self) -> Option<Self::Item> {
-		match &mut self.current {
-			Some(current) => match current.next() {
-				Some(row) => return Some(row),
-				None => {}
-			},
+		// If current is None, iteration is done
+		match self.current.as_mut()?.next() {
+			Some(mut row) => {
+				if let Err(e) = row.as_mut() {
+					e.source = self.index;
+				}
+				return Some(row);
+			}
 			None => {}
 		};
+		// If current was done, go to the next pipeline
 		match self.pipelines.next() {
 			Some(pipeline) => {
-				if pipeline.headers.get_row() != self.headers.get_row() {
-					return Some(Err(Error::MismatchedHeaders(
-						self.headers.get_row().to_owned(),
-						pipeline.headers.get_row().to_owned(),
+				self.index += 1;
+				self.current = Some(pipeline.build());
+				let current = self.current.as_mut().unwrap();
+				if current.headers.get_row() != self.headers.get_row() {
+					return Some(Err(Error::new(
+						self.index,
+						ErrorKind::MismatchedHeaders(
+							self.headers.get_row().to_owned(),
+							current.headers.get_row().to_owned(),
+						),
 					)));
 				}
-				self.current = Some(pipeline.build());
-				self.index += 1;
 			}
 			None => {
 				self.current = None;
+				return None;
 			}
 		}
-		match self.current {
-			Some(ref mut current) => current.next(),
-			None => None,
-		}
+		self.next()
 	}
 }
 
@@ -101,6 +107,7 @@ pub struct MapCol<I, F: FnMut(&str) -> Result<String, Error>> {
 	pub iterator: I,
 	pub f: F,
 	pub name: String,
+	pub source: usize,
 	pub index: Option<usize>,
 }
 impl<I, F> Iterator for MapCol<I, F>
@@ -118,11 +125,21 @@ where
 		let mut row_vec: Vec<_> = row.into_iter().collect();
 		let index = match self.index {
 			Some(index) => index,
-			None => return Some(Err(Error::MissingColumn(self.name.clone()))),
+			None => {
+				return Some(Err(Error::new(
+					self.source,
+					ErrorKind::MissingColumn(self.name.clone()),
+				)))
+			}
 		};
 		let field = match row_vec.get_mut(index) {
 			Some(field) => field,
-			None => return Some(Err(Error::MissingColumn(self.name.clone()))),
+			None => {
+				return Some(Err(Error::new(
+					self.source,
+					ErrorKind::MissingColumn(self.name.clone()),
+				)))
+			}
 		};
 		let new_value = match (self.f)(field) {
 			Ok(value) => value,
@@ -136,6 +153,7 @@ where
 pub struct Select<I> {
 	pub iterator: I,
 	pub columns: Vec<String>,
+	pub source: usize,
 	pub headers: Headers,
 }
 impl<I> Iterator for Select<I>
@@ -153,7 +171,12 @@ where
 		for col in &self.columns {
 			let field = match self.headers.get_field(&row, col) {
 				Some(field) => field,
-				None => return Some(Err(Error::MissingColumn(col.clone()))),
+				None => {
+					return Some(Err(Error::new(
+						self.source,
+						ErrorKind::MissingColumn(col.clone()),
+					)))
+				}
 			};
 			selection.push(field);
 		}
@@ -169,6 +192,7 @@ where
 	pub groups: LinkedHashMap<u64, Vec<Box<dyn Transform>>>,
 	pub hashers: Vec<Box<dyn Transform>>,
 	pub get_transformers: F,
+	pub source: usize,
 	pub headers: Headers,
 }
 impl<I, F> Iterator for TransformInto<I, F>
@@ -188,7 +212,7 @@ where
 			};
 			let hash = match compute_hash(&self.hashers, &self.headers, &row) {
 				Ok(hash) => hash,
-				Err(e) => return Some(Err(e)),
+				Err(e) => return Some(Err(Error::new(self.source, e))),
 			};
 
 			match self.groups.entry(hash) {
@@ -203,7 +227,7 @@ where
 			for reducer in group_row {
 				let result = reducer.add_row(&self.headers, &row);
 				if let Err(e) = result {
-					return Some(Err(e));
+					return Some(Err(Error::new(self.source, e)));
 				}
 			}
 		}
@@ -247,6 +271,7 @@ pub struct ValidateCol<I, F> {
 	pub name: String,
 	pub iterator: I,
 	pub f: F,
+	pub source: usize,
 	pub headers: Headers,
 }
 impl<I, F> Iterator for ValidateCol<I, F>
@@ -263,7 +288,12 @@ where
 		};
 		let field = match self.headers.get_field(&row, &self.name) {
 			Some(field) => field,
-			None => return Some(Err(Error::MissingColumn(self.name.clone()))),
+			None => {
+				return Some(Err(Error::new(
+					self.source,
+					ErrorKind::MissingColumn(self.name.clone()),
+				)))
+			}
 		};
 		match (self.f)(&field) {
 			Ok(()) => Some(Ok(row)),
@@ -275,14 +305,16 @@ where
 pub struct Flush<I, T> {
 	pub iterator: I,
 	pub target: T,
+	pub source: usize,
 	/// `None` if headers have been written, `Some` otherwise
 	headers: Option<Headers>,
 }
 impl<I, T> Flush<I, T> {
-	pub fn new(iterator: I, target: T, headers: Headers) -> Self {
+	pub fn new(iterator: I, target: T, source: usize, headers: Headers) -> Self {
 		Self {
 			iterator,
 			target,
+			source,
 			headers: Some(headers),
 		}
 	}
@@ -298,7 +330,7 @@ where
 		if let Some(headers) = &self.headers {
 			match self.target.write_headers(headers) {
 				Ok(()) => self.headers = None,
-				Err(e) => return Some(Err(e)),
+				Err(e) => return Some(Err(Error::new(self.source, ErrorKind::Csv(e)))),
 			}
 		}
 
@@ -308,7 +340,7 @@ where
 		};
 		let r = match self.target.write_row(&row) {
 			Ok(()) => Some(Ok(row)),
-			Err(e) => Some(Err(e)),
+			Err(e) => return Some(Err(Error::new(self.source, ErrorKind::Csv(e)))),
 		};
 		r
 	}