+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add Pipeline select method
 - Add Pipeline `from_pipelines` constructor for merging pipelines together
 - Remember row order in transform_into
+- Include source index in errors
 
 ## 0.2.0 - 2023 Jan 11
 - Add `Target` struct helper for creating targets, and hide the targets in the `target` module.